fix: attempt to handle indentation of rules and include directives
this is the first attempt at adding support for conditional blocks of rules and include directives.
this version does not evaluate the python content, but rather attempts to simply detect it and pass
the indentation along to the output. this introduces an issue that python/snakemake doesn't
encounter: conditional inclusion may end up presenting the possibility of multiple rules with the
same name. at the moment, the code has no way of distinguishing between them. anticipating possible
issues, I will look to implement a simple check that makes sure duplicate rules are identical in
content. this issue will eventually be solved by python integration.

re #3 #5 #10 #11