fix: address global wildcard_constraints block
recognizes global wildcard_constraints and hides it from the interpreter, but still reports it to
the final synthetic snakefile. this will need to be extended for each additional snakemake feature
that is supported. may consider extending the regex to be more aggressive about detecting floating
blocks of this syntax.

re #25