test: update tests to reflect upcoming changes to inclusion behavior
with the change to python/snakemake integration, it's now possible to keep include directives in
files as-is, and this saves a whole bunch of downstream problems. this patch modifies the tests
involving include directives to expect the new output format (tentatively). the output test status
is not changed to XFAIL as I expect to patch this behavior immediately.

BREAKING CHANGE: tests with include statements now fail, and until later patches are implicitly
XFAIL

re #13 #18 #20 #21 #25