feat: add exclusion paths and extensions, and byte comparison extensions, to config
this information is now received in the input config and emitted to the output tracking config. it
can be picked up by pytest there. these options are not currently supported on the command line, and
I'm not entirely sure whether I want them to be. I'll think about it. note that because the config
tracker is omitted from TAP test comparisons, no changes are needed in the CI tests at this time

re #16 #18 #28