feat: support checkpoints; support certain rules. structures
checkpoints should now work. dependent upstream rules are somewhat overaggressively added to the
synthetic snakefiles, to allow the dag to resolve. for timestamp reasons, the first pass through a
pytest run may execute more than just the single target rule. this may potentially be extremely
time-consuming in some instances; I'm still debating adding a --touch pass to inst/test.py to try to
work around this. note that --touch doesn't interact nicely with directory() flagged output. this
also patches in rules. support with similar logic, though it only currently works when rules.
declarations are present explicitly in a rule block (as opposed to aliased under a function or
something similar). planning to patch in support for a flag to override default behavior and add in
additional dependencies to force rules to work, even when the parser logic can't figure out what is
required.

re #18 #22 #29