feat: add --include-entire-dag option for emergency situations
specifically: 'rules.' support is not guaranteed to work. in such cases, rather than filing a bug
report or giving up, the user can choose to specify this option, and have all rules emitted to the
output test snakefile. this will slow things down significantly in some test cases, and should only
be used sparingly and as a last resort