fix: further repair behavior for orphaned checkpoints
previous fix added partial support but didn't deal with the situation where the orphaned checkpoints
emit output which is then not present in the expected output content. now the output for each
instance of a rule in the input log is emitted to an exclusion list that is used by the pytest
infrastructure to skip expected aberrant output files. these output files are not compared here, but
the checkpoint rules themselves should have their own unit tests, so hopefully this does not
introduce any problems.