fix: remove bugs and generally improve handling of checkpoints
previously the checkpoint handling logic incorrectly included a chain of rules that were completely
unnecessary for actual rule execution, and in some cases were causing additional rules to execute
during test runs, which then caused pytest to crash with errors due to unexpected intermediate
output files. this logic is now much improved: if a rule is updated by a checkpoint, that rule is
included as usual, but then also any checkpoint in the entire pipeline is included in the synthetic
snakefile to address worst case behavior that a rule is incidentally pulling in checkpoints that are
not implicated in the traceable DAG. this also includes a small patch for derived rule resolution
that cleans up the synthetic snakefile and avoids certain weird bugs if the derived rules are buried
under python logic.