fix: replace deprecated function with new lexical parser
bugs in previous commit were trivial issues with whitespace and blank lines, this thing is working.
tests are updated to reflect new, improved output format, and catch whitespace/blank line issues.
only quirky issue now is that string literals are now only being skipped when they're rule
docstrings, which is strange and likely completely unnecessary. this is an easy fix in code due to
the lexical parser, but it'll be tedious to go back to the TAP tests and adjust the output, so for
the moment I'm leaving it.

re #15