test: add unit tests for rule_block primary snakefile parsing logic
since the control for parsing snakefile content is split between two embedded functions, the tests
probe the top-level function directly. there is a small change in the interface for the functions:
once, the name of the file being parsed was passed in for logging reasons, but that seems to not
have survived to this current version of the code, and the filename was being passed around for no
reason.

re #40