feat: draft support for include-rules specification in configuration
at long last, to coexist alongside exclude-rules, a list of rules to be explicitly permitted in the
output unit tests can be provided on the command line or in the config. multiple specification rules
are the same as those for exclude-rules. if both lists are specified, a rule in the include list
will only be included if it is also not in the exclude list.

re #21