feat: change file exclusion and byte comparison specification in pytest to regular expressions
exclude paths and extensions are now condensed into a single exclude-patterns config block, which
accepts regular expressions that are searched against the path and filename of a candidate pytest
file. TAP test for minimum functionality of the exclusion pattern matching is added.

BREAKING CHANGE: old config blocks for exclusion specification are collapsed into a single
`exclude-paths` block; this will need to be manually migrated for existing config files.

re #60