fix: have iterative dryruns respect execution directory
the mid-run dryruns to detect checkpoints and rules. usage was not using the same snakemake call
syntax as the pytest infrastructure, which was a silent issue until tested with a pipeline that is
not executed in its own top-level directory. need to additionally add some sort of check that
determines if an error other than the checkpoint/rules. error is emitted by the dryrun